+++ crates/cargo-contract/src/cmd/extrinsics/remove.rs
@@ -15,13 +15,8 @@
 // along with cargo-contract.  If not, see <http://www.gnu.org/licenses/>.
 
 use super::{
-    runtime_api::api::{
-        self,
-        contracts::events::CodeRemoved,
-    },
     submit_extrinsic,
     Client,
-    CodeHash,
     ContractMessageTranscoder,
     DefaultConfig,
     ExtrinsicOpts,
@@ -29,10 +24,17 @@ use super::{
     TokenMetadata,
 };
 use crate::{
-    cmd::extrinsics::{
-        events::DisplayEvents,
-        parse_code_hash,
-        ErrorVariant,
+    cmd::{
+        extrinsics::{
+            events::DisplayEvents,
+            parse_code_hash,
+            ErrorVariant,
+        },
+        runtime_api::api::{
+            self,
+            contracts::events::CodeRemoved,
+        },
+        CodeHash,
     },
     name_value_println,
 };
@@ -51,7 +53,7 @@ pub struct RemoveCommand {
     code_hash: Option<<DefaultConfig as Config>::Hash>,
     #[clap(flatten)]
     extrinsic_opts: ExtrinsicOpts,
-    /// Export the call output in JSON format.
+    /// Export the call output as JSON.
     #[clap(long, conflicts_with = "verbose")]
     output_json: bool,
 }