+++ crates/cargo-contract/src/cmd/extrinsics/upload.rs
@@ -15,25 +15,28 @@
 // along with cargo-contract.  If not, see <http://www.gnu.org/licenses/>.
 
 use super::{
-    runtime_api::api::{
-        self,
-        runtime_types::pallet_contracts::wasm::Determinism,
-    },
+    display_dry_run_result_warning,
     state_call,
     submit_extrinsic,
-    Balance,
     Client,
-    CodeHash,
     DefaultConfig,
     ExtrinsicOpts,
     PairSigner,
     TokenMetadata,
 };
 use crate::{
-    cmd::extrinsics::{
-        events::DisplayEvents,
-        ErrorVariant,
-        WasmCode,
+    cmd::{
+        extrinsics::{
+            events::DisplayEvents,
+            ErrorVariant,
+            WasmCode,
+        },
+        runtime_api::api::{
+            self,
+            runtime_types::pallet_contracts::wasm::Determinism,
+        },
+        Balance,
+        CodeHash,
     },
     name_value_println,
 };
@@ -78,7 +81,7 @@ impl UploadCommand {
             let url = self.extrinsic_opts.url_to_string();
             let client = OnlineClient::from_url(url.clone()).await?;
 
-            if self.extrinsic_opts.dry_run {
+            if !self.extrinsic_opts.execute {
                 match self.upload_code_rpc(code, &client, &signer).await? {
                     Ok(result) => {
                         let upload_result = UploadDryRunResult {
@@ -90,6 +93,7 @@ impl UploadCommand {
                             println!("{}", upload_result.to_json()?);
                         } else {
                             upload_result.print();
+                            display_dry_run_result_warning("upload");
                         }
                     }
                     Err(err) => {
@@ -102,7 +106,6 @@ impl UploadCommand {
                         }
                     }
                 }
-                Ok(())
             } else if let Some(code_stored) =
                 self.upload_code(&client, code, &signer).await?
             {
@@ -114,14 +117,14 @@ impl UploadCommand {
                 } else {
                     upload_result.print();
                 }
-                Ok(())
             } else {
                 let code_hash = hex::encode(code_hash);
-                Err(anyhow::anyhow!(
+                return Err(anyhow::anyhow!(
                     "This contract has already been uploaded with code hash: 0x{code_hash}"
                 )
                 .into())
             }
+            Ok(())
         })
     }
 
@@ -157,7 +160,7 @@ impl UploadCommand {
         let token_metadata = TokenMetadata::query(client).await?;
         let storage_deposit_limit =
             self.extrinsic_opts.storage_deposit_limit(&token_metadata)?;
-        let call = super::runtime_api::api::tx().contracts().upload_code(
+        let call = crate::cmd::runtime_api::api::tx().contracts().upload_code(
             code.0,
             storage_deposit_limit,
             Determinism::Deterministic,