+++ crates/cargo-contract/src/cmd/mod.rs
@@ -16,6 +16,9 @@
 
 pub mod build;
 pub mod decode;
+pub mod encode;
+pub mod info;
+pub mod runtime_api;
 
 pub(crate) use self::{
     build::{
@@ -23,6 +26,7 @@ pub(crate) use self::{
         CheckCommand,
     },
     decode::DecodeCommand,
+    info::InfoCommand,
 };
 mod extrinsics;
 
@@ -33,3 +37,13 @@ pub(crate) use self::extrinsics::{
     RemoveCommand,
     UploadCommand,
 };
+
+use subxt::{
+    Config,
+    OnlineClient,
+};
+
+pub use subxt::PolkadotConfig as DefaultConfig;
+type Client = OnlineClient<DefaultConfig>;
+type Balance = u128;
+type CodeHash = <DefaultConfig as Config>::Hash;