+++ crates/cargo-contract/src/cmd/runtime_api/mod.rs
@@ -17,7 +17,7 @@
 #![allow(clippy::too_many_arguments)]
 
 #[subxt::subxt(
-    runtime_metadata_path = "src/cmd/extrinsics/runtime_api/contracts_runtime.scale",
+    runtime_metadata_path = "src/cmd/runtime_api/contracts_runtime.scale",
     substitute_type(
         type = "sp_weights::weight_v2::Weight",
         with = "::sp_weights::Weight"