+++ crates/cargo-contract/tests/decode.rs
@@ -74,6 +74,7 @@ fn decode_works() {
     tracing::debug!("Building contract in {}", project_dir.to_string_lossy());
     cargo_contract(&project_dir).arg("build").assert().success();
 
+    // when
     let msg_data: &str = "babebabe01";
     let msg_decoded: &str = r#"switch { value: true }"#;
 
@@ -89,10 +90,27 @@ fn decode_works() {
         .success()
         .stdout(predicates::str::contains(msg_decoded));
 
+    // and when
+    let wrong_msg_data: &str = "babebabe010A";
+    let error_msg: &str = "input length was longer than expected by 1 byte(s).\nManaged to decode `switch`, `value` but `0A` bytes were left unread";
+
+    // then
+    // wrong message data is being handled properly
+    cargo_contract(&project_dir)
+        .arg("decode")
+        .arg("--data")
+        .arg(wrong_msg_data)
+        .arg("-t")
+        .arg("message")
+        .assert()
+        .failure()
+        .stderr(predicates::str::contains(error_msg));
+
+    // when
     let event_data: &str = "080001";
     let event_decoded: &str = r#"Switched { new_value: true }"#;
 
-    // and
+    // then
     // event data is being decoded properly
     cargo_contract(&project_dir)
         .arg("decode")
@@ -104,11 +122,28 @@ fn decode_works() {
         .success()
         .stdout(predicates::str::contains(event_decoded));
 
+    // and when
+    let wrong_event_data: &str = "0800010C";
+    let error_msg: &str = "input length was longer than expected by 1 byte(s).\nManaged to decode `Switched`, `new_value` but `0C` bytes were left unread";
+
+    // then
+    // wrong event data is being handled properly
+    cargo_contract(&project_dir)
+        .arg("decode")
+        .arg("--data")
+        .arg(wrong_event_data)
+        .arg("-t")
+        .arg("event")
+        .assert()
+        .failure()
+        .stderr(predicates::str::contains(error_msg));
+
+    // when
     let constructor_data: &str = "babebabe00";
     let constructor_decoded: &str = r#"new { init_value: false }"#;
 
-    // and
-    // event data is being decoded properly
+    // then
+    // constructor data is being decoded properly
     cargo_contract(&project_dir)
         .arg("decode")
         .arg("-d")
@@ -118,4 +153,20 @@ fn decode_works() {
         .assert()
         .success()
         .stdout(predicates::str::contains(constructor_decoded));
+
+    // and when
+    let wrong_constructor_data: &str = "babebabe00AC";
+    let error_msg: &str = "input length was longer than expected by 1 byte(s).\nManaged to decode `new`, `init_value` but `AC` bytes were left unread";
+
+    // then
+    // wrong constructor data is being handled properly
+    cargo_contract(&project_dir)
+        .arg("decode")
+        .arg("-d")
+        .arg(wrong_constructor_data)
+        .arg("-t")
+        .arg("constructor")
+        .assert()
+        .failure()
+        .stderr(predicates::str::contains(error_msg));
 }