+++ crates/transcode/src/lib.rs
@@ -96,6 +96,7 @@
 //! println!("Encoded constructor data {:?}", data);
 //! ```
 
+mod account_id;
 mod decode;
 mod encode;
 pub mod env_types;
@@ -104,6 +105,7 @@ mod transcoder;
 mod util;
 
 pub use self::{
+    account_id::AccountId32,
     scon::{
         Hex,
         Map,
@@ -277,6 +279,8 @@ impl ContractMessageTranscoder {
             args.push((Value::String(name), value));
         }
 
+        Self::validate_length(data, event_spec.label(), &args)?;
+
         let name = event_spec.label().to_string();
         let map = Map::new(Some(&name), args.into_iter().collect());
 
@@ -292,7 +296,7 @@ impl ContractMessageTranscoder {
             .ok_or_else(|| {
                 anyhow::anyhow!(
                     "Message with selector {} not found in contract metadata",
-                    hex::encode(msg_selector)
+                    hex::encode_upper(msg_selector)
                 )
             })?;
         tracing::debug!("Decoding contract message '{}'", msg_spec.label());
@@ -303,6 +307,8 @@ impl ContractMessageTranscoder {
             let value = self.decode(arg.ty().ty().id(), data)?;
             args.push((Value::String(name), value));
         }
+
+        Self::validate_length(data, msg_spec.label(), &args)?;
 
         let name = msg_spec.label().to_string();
         let map = Map::new(Some(&name), args.into_iter().collect());
@@ -319,7 +325,7 @@ impl ContractMessageTranscoder {
             .ok_or_else(|| {
                 anyhow::anyhow!(
                     "Constructor with selector {} not found in contract metadata",
-                    hex::encode(msg_selector)
+                    hex::encode_upper(msg_selector)
                 )
             })?;
         tracing::debug!("Decoding contract constructor '{}'", msg_spec.label());
@@ -330,6 +336,8 @@ impl ContractMessageTranscoder {
             let value = self.decode(arg.ty().ty().id(), data)?;
             args.push((Value::String(name), value));
         }
+
+        Self::validate_length(data, msg_spec.label(), &args)?;
 
         let name = msg_spec.label().to_string();
         let map = Map::new(Some(&name), args.into_iter().collect());
@@ -346,6 +354,24 @@ impl ContractMessageTranscoder {
         } else {
             Ok(Value::Unit)
         }
+    }
+
+    /// Checks if buffer empty, otherwise returns am error
+    fn validate_length(data: &[u8], label: &str, args: &[(Value, Value)]) -> Result<()> {
+        if !data.is_empty() {
+            let arg_list_string: String =
+                args.iter().fold(format!("`{label}`"), |init, arg| {
+                    format!("{}, `{}`", init, arg.0)
+                });
+            let encoded_bytes = hex::encode_upper(data);
+            return Err(anyhow::anyhow!(
+                "input length was longer than expected by {} byte(s).\nManaged to decode {} but `{}` bytes were left unread",
+                data.len(),
+                arg_list_string,
+                encoded_bytes
+            ))
+        }
+        Ok(())
     }
 }
 
@@ -545,10 +571,9 @@ mod tests {
         // encoded args follow the 4 byte selector
         let encoded_args = &encoded[4..];
 
-        let expected = sp_core::crypto::AccountId32::from_str(
-            "5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY",
-        )
-        .unwrap();
+        let expected =
+            AccountId32::from_str("5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY")
+                .unwrap();
         assert_eq!(expected.encode(), encoded_args);
         Ok(())
     }
@@ -567,14 +592,10 @@ mod tests {
         let encoded_args = &encoded[4..];
 
         let expected = vec![
-            sp_core::crypto::AccountId32::from_str(
-                "5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY",
-            )
-            .unwrap(),
-            sp_core::crypto::AccountId32::from_str(
-                "5FHneW46xGXgs5mUiveU4sbTyGBzmstUspZC92UhjJM694ty",
-            )
-            .unwrap(),
+            AccountId32::from_str("5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY")
+                .unwrap(),
+            AccountId32::from_str("5FHneW46xGXgs5mUiveU4sbTyGBzmstUspZC92UhjJM694ty")
+                .unwrap(),
         ];
         assert_eq!(expected.encode(), encoded_args);
         Ok(())
@@ -688,7 +709,7 @@ mod tests {
         let transcoder = ContractMessageTranscoder::new(metadata);
 
         // raw encoded event with event index prefix
-        let encoded = (0u8, [0u32; 32], [1u32; 32]).encode();
+        let encoded = (0u8, [0u32; 8], [1u32; 8]).encode();
         // encode again as a Vec<u8> which has a len prefix.
         let encoded_bytes = encoded.encode();
         let _ = transcoder.decode_contract_event(&mut &encoded_bytes[..])?;
@@ -706,7 +727,7 @@ mod tests {
             83, 118, 135, 56, 220, 172, 95, 131, 171, 125, 130, 167, 10, 15, 242, 222,
         ];
         // raw encoded event with event index prefix
-        let encoded = (0u8, hash, [0u32; 32]).encode();
+        let encoded = (0u8, hash, [0u32; 8]).encode();
         // encode again as a Vec<u8> which has a len prefix.
         let encoded_bytes = encoded.encode();
         let decoded = transcoder.decode_contract_event(&mut &encoded_bytes[..])?;
@@ -738,5 +759,36 @@ mod tests {
         let _ = transcoder.decode_contract_message(&mut &encoded_bytes[..])?;
 
         Ok(())
+    }
+
+    #[test]
+    #[should_panic(
+        expected = "input length was longer than expected by 1 byte(s).\nManaged to decode `flip` but `00` bytes were left unread"
+    )]
+    fn fail_decode_input_with_extra_bytes() {
+        let metadata = generate_metadata();
+        let transcoder = ContractMessageTranscoder::new(metadata);
+
+        let encoded_bytes = hex::decode("633aa55100").unwrap();
+        let _ = transcoder
+            .decode_contract_message(&mut &encoded_bytes[..])
+            .unwrap();
+    }
+
+    #[test]
+    #[should_panic(
+        expected = "input length was longer than expected by 2 byte(s).\nManaged to decode `Event1`, `name`, `from` but `0C10` bytes were left unread"
+    )]
+    fn fail_decode_contract_event_with_extra_bytes() {
+        let metadata = generate_metadata();
+        let transcoder = ContractMessageTranscoder::new(metadata);
+
+        // raw encoded event with event index prefix
+        let encoded = (0u8, [0u32; 8], [1u32; 8], [12u8, 16u8]).encode();
+        // encode again as a Vec<u8> which has a len prefix.
+        let encoded_bytes = encoded.encode();
+        let _ = transcoder
+            .decode_contract_event(&mut &encoded_bytes[..])
+            .unwrap();
     }
 }